Remove old unused code
