--- src/createActionId.ts
@@ -1,5 +1,0 @@
-import { getGlobalContext } from './globalContext';
-
-export default function createActionId(): string {
-    return (getGlobalContext().nextActionId++).toString();
-}
