+++ src/index.ts
@@ -7,13 +7,10 @@ export { default as DispatchFunction } from './interfaces/DispatchFunction';
 export { default as Middleware } from './interfaces/Middleware';
 export { default as MutatorFunction } from './interfaces/MutatorFunction';
 export { default as OrchestratorFunction } from './interfaces/OrchestratorFunction';
-export { action, actionCreator } from './actionCreator';
-export { default as applyMiddleware } from './applyMiddleware';
-export { default as createStore } from './createStore';
-export { dispatch } from './dispatcher';
+export { Mutator } from './interfaces/Mutator';
+export { Orchestrator } from './interfaces/Orchestrator';
 export { default as mutator } from './mutator';
 import { default as orchestrator } from './orchestrator';
-export { default as getRootStore } from './getRootStore';
 export { mutatorAction } from './simpleSubscribers';
 export { useStrict };
