+++ src/interfaces/Mutator.ts
@@ -6,5 +6,4 @@ export type Mutator<TAction extends ActionMessage, TReturn> = {
     type: 'mutator';
     actionCreator: ActionCreator<TAction>;
     target: MutatorFunction<TAction, TReturn>;
-    isRegistered: boolean;
 };
