+++ src/simpleSubscribers.ts
@@ -1,23 +1,31 @@
 import SimpleAction from './interfaces/SimpleAction';
-import { action } from './actionCreator';
 import mutator from './mutator';
+import { SatchelInstance } from './createSatchel';
 
 export function createSimpleSubscriber(decorator: Function) {
     return function simpleSubscriber<TFunction extends (...args: any) => any>(
+        satchelInstance: SatchelInstance,
         actionType: string,
         target: TFunction
     ): SimpleAction<TFunction> {
         // Create the action creator
-        let simpleActionCreator = action(actionType, function simpleActionCreator() {
-            return {
-                args: arguments,
-            };
-        });
+        let simpleActionCreator = satchelInstance.action(
+            actionType,
+            function simpleActionCreator() {
+                return {
+                    args: arguments,
+                };
+            }
+        );
 
         // Create the subscriber
-        decorator(simpleActionCreator, function simpleSubscriberCallback(actionMessage: any) {
+        const subscriber = decorator(simpleActionCreator, function simpleSubscriberCallback(
+            actionMessage: any
+        ) {
             return target.apply(null, actionMessage.args);
         });
+
+        satchelInstance.register(subscriber);
 
         // Return a function that dispatches that action
         return (simpleActionCreator as any) as SimpleAction<TFunction>;
